refactor(helm): rename client wrapper consistently
