+++ kpops/component_handlers/helm/exception.py
@@ -1,6 +1,6 @@
 from typing import ClassVar
 
-from kpops.component_handlers.helm_wrapper import HELM
+from kpops.component_handlers.helm import HELM
 from kpops.core.exception import ServiceException
