+++ kpops/components/streams_bootstrap/base.py
@@ -9,7 +9,7 @@ import structlog
 from pydantic import Field
 from typing_extensions import override
 
-from kpops.component_handlers.helm_wrapper.model import HelmRepoConfig
+from kpops.component_handlers.helm.model import HelmRepoConfig
 from kpops.components.base_components import KafkaApp
 from kpops.components.base_components.cleaner import Cleaner
 from kpops.components.base_components.helm_app import HelmApp
