+++ kpops/config/__init__.py
@@ -15,7 +15,7 @@ from pydantic_settings import (
 )
 from typing_extensions import override
 
-from kpops.component_handlers.helm_wrapper.model import HelmConfig
+from kpops.component_handlers.helm.model import HelmConfig
 from kpops.core.exception import ValidationError
 from kpops.core.operation import OperationMode
 from kpops.utils.docstring import describe_object
