+++ tests/components/streams_bootstrap_v2/test_streams_bootstrap.py
@@ -4,11 +4,11 @@ import pytest
 from pydantic import ValidationError
 from pytest_mock import MockerFixture
 
-from kpops.component_handlers.helm_wrapper.model import (
+from kpops.component_handlers.helm.model import (
     HelmRepoConfig,
     HelmUpgradeInstallFlags,
 )
-from kpops.component_handlers.helm_wrapper.utils import create_helm_release_name
+from kpops.component_handlers.helm.utils import create_helm_release_name
 from kpops.components.streams_bootstrap_v2 import StreamsBootstrapV2
 from kpops.components.streams_bootstrap_v2.base import StreamsBootstrapV2Values
