+++ tests/pipeline/test_manifest.py
@@ -10,8 +10,8 @@ from typer.testing import CliRunner
 
 import kpops.api as kpops
 from kpops.cli.main import app
-from kpops.component_handlers.helm_wrapper.helm import Helm
-from kpops.component_handlers.helm_wrapper.model import HelmConfig, Version
+from kpops.component_handlers.helm.helm import Helm
+from kpops.component_handlers.helm.model import HelmConfig, Version
 from kpops.const.file_type import PIPELINE_YAML
 from kpops.manifests.kubernetes import KubernetesManifest
 from kpops.utils.yaml import print_yaml
